Module Toml__Printer

val value : Stdlib.Format.formatter -> Toml.Types.value -> unit

Given a Toml value and a formatter, inserts a valid Toml representation of this value in the formatter.

since

2.0.0

val table : Stdlib.Format.formatter -> Toml.Types.table -> unit

Given a Toml table and a formatter, inserts a valid Toml representation of this value in the formatter.

since

2.0.0

val array : Stdlib.Format.formatter -> Toml.Types.array -> unit

Given a Toml array and a formatter, inserts a valid Toml representation of this value in the formatter.

raises Invalid_argument

if the array is an array of tables

since

2.0.0

val string_of_value : Toml.Types.value -> string

Turns a Toml value into a string.

since

4.0.0

val string_of_table : Toml.Types.table -> string

Turns a Toml table into a string.

since

4.0.0

val string_of_array : Toml.Types.array -> string

Turns a Toml array into a string.

since

4.0.0